+++ examples/Web Apps/Framework Examples/React/ServiceAccountManager/src/app/ServiceAccountManager.tsx
@@ -1,5 +1,5 @@
 import { type JSX } from 'react';
-import { UserServiceTable } from './components/UserServiceTable';
+import { UserServiceTable } from './components/ServiceAccountTable';
 
 const ServiceAccountManager = (): JSX.Element => {
     return (
